CONFIG: Make loop configurable
