+++ misfit_toys/examples/source2/cfg/train/full.yaml
@@ -5,3 +5,4 @@ defaults:
 
 n_epochs: 100
 lr: 0.1
+loop: ^^cwd|helpers|vanilla_train
